+++ lib/services/isar_provider.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:isar/isar.dart';
-import 'package:padelistas/services/settings/settings.dart';
 import 'package:path_provider/path_provider.dart';
+
+import 'settings/settings.dart';
 
 final isarProvider = FutureProvider<Isar>((ref) async {
   final dir = kIsWeb